requirements.txt
python-dotenv
scapy